requirements.txt
python-dotenv==0.19.2
python_telegram_bot==13.9
requests==2.25.1
telegram==0.0.1
